backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.4
pydantic-settings==2.3.0
langgraph==0.1.19
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-community==0.2.5
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
pgvector==0.2.5
redis[hiredis]==5.0.4
httpx==0.27.0
GitPython==3.1.43
PyGithub==2.3.0
python-gitlab==4.6.0
cryptography==42.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-client==0.20.0
structlog==24.2.0
docker==7.1.0
python-multipart==0.0.9
python-dotenv==1.0.1
tenacity==8.3.0
anthropic==0.28.0
numpy==1.26.4
